requirements.txt
# Django Core
Django>=4.2,<5.0
django-cors-headers>=4.2,<5.0
dj-database-url>=2.1,<3.0
django-ratelimit>=4.1,<5.0

# Modern Admin UI
django-jazzmin>=3.0,<4.0

# Database
psycopg2-binary>=2.9.9,<3.0

# Async Tasks - Commented out (not currently used)
# celery>=5.3,<6.0
# redis>=4.6,<5.0
# django-celery-beat>=2.5,<3.0
# django-celery-results>=2.5,<3.0

# File Handling
Pillow>=10.0,<11.0
qrcode>=7.4,<8.0

# Production Server
gunicorn>=21.0,<22.0
whitenoise>=6.5,<7.0

# Environment
python-dotenv>=1.0,<2.0

# Utilities
python-dateutil>=2.8,<3.0

# PDF Generation (for receipts)
reportlab>=4.0,<5.0

# Performance Optimization
csscompressor>=0.9.5
jsmin>=3.0.1
# django-redis>=5.4.0  # Not needed - using locmem cache

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7

# Payment Gateways
stripe>=7.0.0
paypalrestsdk>=1.13.1

# Real-time (Django Channels) - Commented out (not currently used)
# channels>=4.0.0
# channels-redis>=4.1.0
# daphne>=4.0.0
